# docs/openapi3.yaml
openapi: 3.0.0
info:
  description: Production-ready backend Endge-сервиса с эталонными
    RedPanda-обёртками и строгими архитектурными границами.
  title: Endge Service Backend API
  contact: {}
  version: 1.0.20
paths:
...
  /api/v1/projects:
    get:
      security:
        - BearerAuth: []
      description: Возвращает список неудаленных проектов.
      tags:
        - projects
      summary: Список проектов
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_project.ProjectsListResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Создает новый проект.
      tags:
        - projects
      summary: Создать проект
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_project.CreateProjectRequest"
        description: Параметры проекта
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_project.ProjectResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}":
    get:
      security:
        - BearerAuth: []
      description: Возвращает проект по identity.
      tags:
        - projects
      summary: Получить проект
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_project.ProjectResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    delete:
      security:
        - BearerAuth: []
      description: Выполняет soft delete проекта по identity.
      tags:
        - projects
      summary: Удалить проект
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    patch:
      security:
        - BearerAuth: []
      description: Обновляет проект по identity.
      tags:
        - projects
      summary: Обновить проект
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_project.UpdateProjectRequest"
        description: Параметры обновления проекта
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_project.ProjectResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
...
  "/api/v1/projects/{project_identity}/hard":
    delete:
      security:
        - BearerAuth: []
      description: Выполняет hard delete проекта по identity.
      tags:
        - projects
      summary: Удалить проект физически
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
...
  "/api/v1/projects/{project_identity}/restore":
    post:
      security:
        - BearerAuth: []
      description: Восстанавливает soft-deleted проект по identity.
      tags:
        - projects
      summary: Восстановить проект
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    internal_api_http_v1_project.CreateProjectRequest:
      type: object
      required:
        - displayName
        - identity
      properties:
        active:
          type: boolean
          example: true
        description:
          type: string
          example: Project for local configuration
        displayName:
          type: string
          maxLength: 255
          minLength: 1
          example: Demo Project
        identity:
          type: string
          maxLength: 160
          minLength: 1
          example: demo-project
        meta:
          type: object
    internal_api_http_v1_project.ProjectResponse:
      type: object
      properties:
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          example: 2026-07-08T10:00:00Z
        deletedAt:
          type: string
          example: 2026-07-08T10:00:00Z
        description:
          type: string
          example: Project for local configuration
        displayName:
          type: string
          example: Demo Project
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        identity:
          type: string
          example: demo-project
        meta:
          type: object
        updatedAt:
          type: string
          example: 2026-07-08T10:00:00Z
    internal_api_http_v1_project.ProjectsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/internal_api_http_v1_project.ProjectResponse"
    internal_api_http_v1_project.UpdateProjectRequest:
      type: object
      required:
        - displayName
      properties:
        active:
          type: boolean
          example: true
        description:
          type: string
          example: Updated description
        displayName:
          type: string
          maxLength: 255
          minLength: 1
          example: Demo Project
        meta:
          type: object
...